Update after previous merge
PR #1443 was merged into this branch to handle all import type in the same place.
