+++ src/Wallabag/CoreBundle/DependencyInjection/Configuration.php
@@ -4,7 +4,6 @@ namespace Wallabag\CoreBundle\DependencyInjection;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
-use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 
 class Configuration implements ConfigurationInterface
 {
@@ -18,21 +17,9 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('languages')
                     ->prototype('scalar')->end()
                 ->end()
-                ->arrayNode('import')
-                    ->append($this->getAllowMimetypes())
-                ->end()
             ->end()
         ;
 
         return $treeBuilder;
-    }
-
-    private function getAllowMimetypes()
-    {
-        $node = new ArrayNodeDefinition('allow_mimetypes');
-
-        $node->prototype('scalar')->end();
-
-        return $node;
     }
 }
