+++ src/Wallabag/CoreBundle/DependencyInjection/WallabagCoreExtension.php
@@ -14,7 +14,6 @@ class WallabagCoreExtension extends Extension
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
         $container->setParameter('wallabag_core.languages', $config['languages']);
-        $container->setParameter('wallabag_core.import', $config['import']);
 
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yml');
